requirements.txt
appdirs==1.4.4
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
grpcio==1.44.0
grpcio-tools==1.44.0
httpstat==1.3.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
locust==1.4.3
MarkupSafe==1.1.1
msgpack==1.0.2
protobuf==3.19.4
psutil==5.8.0
pycodestyle==2.6.0
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
